% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{._bcsimsenv_}
\alias{._bcsimsenv_}
\title{Copyright 2026 Province of British Columbia}
\format{
An object of class \code{environment} of length 0.
}
\usage{
._bcsimsenv_
}
\description{
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
}
\details{
http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
Set package-specific environment & variables
These variables disappear once package is unloaded from user's system
E.g., once user logs on to SIMS, you can store user info into pkg env
for recall in other fxns. Once user is done w bcsims, vars disappear.
https://r-pkgs.org/data.html#'sec-data-state
}
\keyword{datasets}
