% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_sims}
\alias{get_sims}
\title{Copyright 2026 Province of British Columbia}
\usage{
get_sims(...)
}
\arguments{
\item{...}{URL components of the GET request, seperated by commas.}
}
\value{
An object of class 'httr2_request'
}
\description{
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
}
\details{
http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
Functions to build SIMS API GET, POST, and DELETE requests

TODO list:
- consider actually performing the request w `get_sims()` to reduce
   duplicated code bits in every other `get_sims_*` fxn.
Build a SIMS API URL GET request

A utility function to build a valid GET request URL for you. See the SIMS
[sims-api](https://api-biohubbc.apps.silver.devops.gov.bc.ca/api-docs/) documentation for
all valid API endpoints. All other bcsims `get_*` functions are wrappers
around this function.
}
\examples{
# Build URL to grab surveys from project '999':
# /project/{projectId}/survey/
get_sims("project", 999, "survey")

}
