% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{get_sims_devices}
\alias{get_sims_devices}
\title{Get a table of devices within a given project and/or survey}
\usage{
get_sims_devices(project_id, survey_id = NULL)
}
\arguments{
\item{project_id}{A single project ID number.}

\item{survey_id}{Optional. A survey number within a given project. If no `project_id` is supplied, the function will pull devices from all surveys within a given project.}
}
\value{
A tibble of devices.
}
\description{
Get a table of devices within a given project and/or survey
}
\examples{
\dontrun{
get_sims_devices(project_id = 42)

get_sims_devices(project_id = 42, survey_id = 52)
}
}
