% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{get_sims_surveys}
\alias{get_sims_surveys}
\alias{get_sims_surveys.default}
\alias{get_sims_surveys.numeric}
\title{Copyright 2026 Province of British Columbia}
\usage{
get_sims_surveys(...)

\method{get_sims_surveys}{default}(params = NULL, ...)

\method{get_sims_surveys}{numeric}(project_id, ...)
}
\arguments{
\item{...}{Not used.}

\item{params}{A list of API parameters, e.g. `list(keyword = "bctw", keyword = "telemetry", itis_tsns = c(180701, 202411))`. See Details for a list of acceptable parameters. Currently `params` are ignored if a project_id has been supplied.}

\item{project_id}{A single project ID number.}
}
\value{
A tibble of surveys.
}
\description{
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
}
\details{
http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
Get a table of available surveys in SIMS
}
\examples{
\dontrun{
# Get all surveys available to the user
all_surveys <- get_sims_surveys()

# Get all surveys that match a certain criteria
wolf_surveys <- get_sims_surveys(params = list(keyword = "wolf"))
caribou_surveys <- get_sims_surveys(params = list(itis_tsn = 180596))

# Get surveys under a specific project ID
# Note this method ignores the `params` argument
caribou_surveys <- get_sims_surveys(42)
}
}
