% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{get_sims_devices}
\alias{get_sims_devices}
\title{Copyright 2026 Province of British Columbia}
\usage{
get_sims_devices(project_id, survey_id = NULL)
}
\arguments{
\item{project_id}{A single project ID number.}

\item{survey_id}{Optional. A survey number within a given project. If no `project_id` is supplied, the function will pull devices from all surveys within a given project.}
}
\value{
A tibble of devices.
}
\description{
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
}
\details{
http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
Get a table of devices within a given project and/or survey
}
\examples{
\dontrun{
get_sims_devices(project_id = 42)

get_sims_devices(project_id = 42, survey_id = 52)
}
}
